+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSVillagesConfig.java
@@ -14,6 +14,7 @@ public class RSVillagesConfig
 	{
 		public ConfigValueListener<Boolean> addVillagesToModdedBiomes;
 		public ConfigValueListener<Integer> badlandsVillageSpawnrate;
+		public ConfigValueListener<Integer> birchVillageSpawnrate;
 
 		public RSVillagesConfigValues(ForgeConfigSpec.Builder builder, ConfigHelper.Subscriber subscriber)
 		{
@@ -32,6 +33,13 @@ public class RSVillagesConfig
 							+" 1 for spawning in most chunks and 1001 for no spawn.")
         					.translation("repurposedstructures.config.villages.badlandsVillageSpawnrate")
         					.defineInRange("badlandsVillageSpawnrate", 17, 1, 1001));
+				
+					birchVillageSpawnrate = subscriber.subscribe(builder
+						.comment("\r\n How rare are Birch Villages in Birch biomes.\r\n"
+							+"\n "
+							+" 1 for spawning in most chunks and 1001 for no spawn.")
+        					.translation("repurposedstructures.config.villages.birchvillagespawnrate")
+        					.defineInRange("birchVillageSpawnrate", 24, 1, 1001));
 				builder.pop();
 					
 			builder.pop();